+++ frontend/src/app/AppRouter.jsx
@@ -4,6 +4,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Header from "../components/Header.jsx"
 import HomePage from "../pages/HomePage.jsx";
 import NotFoundPage from "../pages/NotFoundPage.jsx";
+import Footer from "../components/Footer.jsx";
 
 export default function AppRouter() {
     return (
@@ -15,6 +16,7 @@ export default function AppRouter() {
                 {/* Page doesn't exist */}
                 <Route path="*" element={<NotFoundPage />} />
             </Routes>
+            <Footer />
         </BrowserRouter>
     );
 }